% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_ref_matern_5_2.R
\name{approx_ref_matern_5_2}
\alias{approx_ref_matern_5_2}
\title{Approximate Mater5_2 kernel}
\usage{
approx_ref_matern_5_2(a, b, C, beta_i, eta_i)
}
\arguments{
\item{a}{parameter.}

\item{b}{parameter.}

\item{C}{parameter.}

\item{beta_i}{the inverse of range parameter.}

\item{eta_i}{the nugget parameter.}
}
\description{
Compute the approximate Matern5_2 kernel.
}
\examples{

approx_ref_matern_5_2(1,1/2,0.1,0.5, 2)

}
\references{
Gu, M., & Xu, Y. (2020). Fast nonseparable Gaussian stochastic process with application to methylation level interpolation. Journal of Computational and Graphical Statistics, 29(2), 250-260.
}
