% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methyl.R
\docType{data}
\name{data_methyl}
\alias{data_methyl}
\title{equineDiffFactors}
\format{
A data frame with 2256 rows and 8 variables
}
\usage{
data(data_methyl)
}
\description{
Equine Influenza dataset
}
\details{
\itemize{
\item ID.source are the ID of source hosts
\item ID.recep are the ID of receiver hosts
  \item y are the vector of transmission probabilities source -> receiver
  \item other columns are the factors
}
}
\references{
\doi{/10.5281/zenodo.4837560}
}
\author{
Melina Ribaud and Joseph Hughes
}
\keyword{equineDiffFactors}
