% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_post.R
\name{log_post}
\alias{log_post}
\title{Log posterior of Matern5_2 kernel}
\usage{
log_post(param, object, C)
}
\arguments{
\item{param}{hyperparameters.}

\item{object}{a FastGaSP object.}

\item{C}{parameter.}
}
\description{
Compute the log posterior of the Matern5_2 kernel.
}
\examples{

sites = sort(runif(100))
Y = sin(2*pi*sites)
C = diff(range(sites))/length(sites)
fgasp.model = fgasp(sites,Y, have_noise = TRUE)
log_post(c(log(1 / C), 1),fgasp.model,C)

}
\references{
Gu, M., & Xu, Y. (2020). Fast nonseparable Gaussian stochastic process with application to methylation level interpolation. Journal of Computational and Graphical Statistics, 29(2), 250-260.
}
