% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern_5_2_kernel.R
\name{matern_5_2_kernel}
\alias{matern_5_2_kernel}
\title{Mater5_2 kernel}
\usage{
matern_5_2_kernel(d, beta)
}
\arguments{
\item{d}{the distance.}

\item{beta}{the inverse of range parameter.}
}
\description{
Compute the matern5_2 kernel
}
\examples{

matern_5_2_kernel(0.5, beta = 2)
 
}
