% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methyl.R
\docType{data}
\name{data_methyl}
\alias{data_methyl}
\title{data_methyl}
\format{
A list with two elements: Y the matrix with the methylation level for each samples and sites,
sites the genomic position of sites,
X the covariate whith the cell type
}
\usage{
data(data_methyl)
}
\description{
Extract from the methylation data set of Ribaud et al. from site 119521761 to 119527907
}
\details{
\itemize{
\item ID.source are the ID of source hosts
\item ID.recep are the ID of receiver hosts
  \item y are the vector of transmission probabilities source -> receiver
  \item other columns are the factors
}
}
\references{
Melina Ribaud, Aurélie Labbe and Karim Oualkacha.
Imputation in genetic methylation studies: A linear model of coregionalization (LMC) with informative covariates.
2022. hal-00000000
}
\author{
Melina Ribaud
}
\keyword{equineDiffFactors}
