% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern_5_2_kernel.R
\name{matern_5_2_kernel}
\alias{matern_5_2_kernel}
\title{Mater5_2 kernel}
\usage{
matern_5_2_kernel(d, beta)
}
\arguments{
\item{d}{the distance.}

\item{beta}{the inverse of range parameter.}
}
\description{
Compute the Matern5_2 kernel.
}
\examples{

matern_5_2_kernel(0.5, beta = 2)

}
\references{
Gu, M., & Xu, Y. (2020).
Fast nonseparable Gaussian stochastic process with application to methylation level interpolation.
Journal of Computational and Graphical Statistics, 29(2), 250-260.
\doi{10.1080/10618600.2019.1665534}
}
