% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsgasp.R
\docType{package}
\name{olsgasp}
\alias{olsgasp}
\title{olsgasp: A package for imputation in genetic methylation studies}
\description{
We propose a a method to predict the missing methylation levels.
This method catches correlation structures among the methylation levels across genome sites and across samples.
The regression function linking the methylation level to the covariates
is modeled through a linear combination of the covariates together with latent factors.
We assume the covariates' and latent factors' effects to be Gaussian random processes (GP).
}
\references{
Melina Ribaud, Aurélie Labbe and Karim Oualkacha.
Imputation in genetic methylation studies: A linear model of coregionalization (LMC) with informative covariates.
2022. hal-00000000
}
\author{
Melina Ribaud
}
